LayoutTests/imported/w3c/web-platform-tests/css/css-transforms/preserve-3d-flat-grouping-properties-containing-block-expected.txt


FAIL Preserve-3d element with a grouping property still a containing block assert_equals: expected Element node <div id="parentWithGrouping" style="transform-style: pres... but got Element node <body><div id="parentWithGrouping" style="transform-style...
FAIL Preserve-3d element without a grouping property still a containing block assert_equals: expected Element node <div id="parentWithoutGrouping" style="transform-style: p... but got Element node <body><div id="parentWithGrouping" style="transform-style...
